Adding a function that generates distinct test sets and runs the policy for each experiment of a specific game. Plots the mean reward for each experiment
Adding a function that gets the policy for each experiment of a specific game, generates a distinct test sets using levels that the models have not seen before. It then plots the test mean reward for each experiment. There is also an option to save test videos for each experiment